Reset text on LabelRidget if we try to set the same text again.
If we try to set the same Text on the LabelRidget, first set Empty
String on the Label Widget and then set the right text. This workaround
is needed because the implementation of the Label widget doesn't set 
the same text again.

Furthermore we save the Image set on the Widget while we bind our Ridget
against the widget.